:truck: refactor change ErrorCode folder
