+++ src/main/java/com/antonio/apprendrebackend/service/controller/HomeController.java
@@ -1,13 +1,11 @@
 package com.antonio.apprendrebackend.service.controller;
 
 
-import com.antonio.apprendrebackend.service.dto.WordTranslationDTO;
 import com.antonio.apprendrebackend.service.model.ErrorResponse;
 import com.antonio.apprendrebackend.service.model.Home;
 import com.antonio.apprendrebackend.service.service.HomeService;
-import com.antonio.apprendrebackend.service.util.ErrorCode;
+import com.antonio.apprendrebackend.service.exception.ErrorCode;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.AutoConfigureOrder;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
