+++ src/main/java/com/antonio/apprendrebackend/service/controller/WordSpController.java
@@ -3,7 +3,7 @@ package com.antonio.apprendrebackend.service.controller;
 import com.antonio.apprendrebackend.service.model.ErrorResponse;
 import com.antonio.apprendrebackend.service.model.WordSp;
 import com.antonio.apprendrebackend.service.repository.WordSpRepository;
-import com.antonio.apprendrebackend.service.util.ErrorCode;
+import com.antonio.apprendrebackend.service.exception.ErrorCode;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
