+++ src/main/java/com/antonio/apprendrebackend/service/exception/ErrorCode.java
@@ -1,0 +1,17 @@
+package com.antonio.apprendrebackend.service.exception;
+
+public enum ErrorCode {
+    TYPE_NOT_FOUND,
+    WORD_FR_EXISTS,
+    INVALID_REQUEST,
+    TYPE_NOT_EXIST,
+    WORD_ES_EXISTS,
+    WORD_TRANSLATION_NOT_FOUND,
+    PHRASE_NOT_FOUND,
+    HOME_NOT_FOUND,
+    GOAL_NOT_CREATED,
+    USER_INFO_NOT_FOUND,
+    USER_REQUEST_LAST_WEEK_NOT_ANSWERED,
+    WORD_TRANSLATION_HISTORIAL_NOT_FOUND,
+    DECK_WORD_TRANSLATION_HISTORIAL_NOT_FOUND
+}
